+++ src/RootComponents.js
@@ -7,7 +7,7 @@ export class DefaultDocument extends Component {
       <Html lang="en-US">
         <Head>
           <meta charSet="UTF-8" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
+          <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
         </Head>
         <Body>{children}</Body>
       </Html>
@@ -21,8 +21,4 @@ export const Html = ({ children, ...rest }) => (
   </html>
 )
 export const Head = ({ children, ...rest }) => <head {...rest}>{children}</head>
-export const Body = ({ children, ...rest }) => (
-  <body {...rest}>
-    {children}
-  </body>
-)
+export const Body = ({ children, ...rest }) => <body {...rest}>{children}</body>